+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 ## 1.0.0
 
 * Initial release
-    - Implemented collapsable app bar when search bar is focused.
-    - Implemented expand and shrink animation of search bar.
+    - Fix default color